Energy system includes class demand_profile.estimate Demand is automatically estimated when energysystem is simulated Results mirror expected values - pv generation much, much higher than needed, generator does not have to be used at all! Next step: include class pvlib_generation
